infra: Move DNS zone management to the separate module
We need to manage Route53 zone separately because it requires manual
configuartion change of NS records in the "Registered Domain" section
and it also takes 2 days to propagate these changes. Meanwhile DNS for
the domain doesn't work.

In this commit we move it to the separate Terraform module and set
`prevent_destroy = true` to avoid accidental destroy. Now in the root
module "infra" we can safely run `terraform destroy` without breaking
DNS.