infra: Add app deploy via ansible
To deploy an app we are using docker-compose that just launches public
images. Ansible is automating the whole process - it prepares the node
by installing docker and docker-compose together with all needed
dependencies, copies docker-compose.yml file and launches it.

To communicate with the EC2 instance dynamic inventory is used. Its
configuration isin inventroy.aws_ec2.yml and it's needed to create
groups from AWS tags. The tags are used as hosts in the main site.yml
playbook.